Make useable versions of message datatypes
Made datatypes corresponding to the ones with the protobuf serialisation
derived, but using normal types rather than leaving it as it comes out
of the decoding, so that it's more convenient to actually use them. This
is done via an overcomplicated Generics mechanism.

Also corrected error code for invalid proof fom 404 to 400.